requirements.txt
torch>=1.8.0
scgpt
numpy>=1.21.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
umap-learn>=0.5.0 